backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.36
asyncpg==0.29.0
aiosqlite==0.20.0
python-dotenv==1.0.1
httpx==0.27.2
openai>=1.54.0
